+++ CHANGELOG.md
@@ -27,10 +27,13 @@
 - `general`
   - [#3728](https://github.com/wix/ricos/pull/3728) RicosEditor API types moved to ricos-types
 - `ricos-editor`
+
   - [#3728](https://github.com/wix/ricos/pull/3728) individual reconfiguration of extensions according to RicosEditor props
   - [#3844](https://github.com/wix/ricos/pull/3844) tiptap plugins config merger
+  - [#3843](https://github.com/wix/ricos/pull/3843) tiptap theme strategy
 
 - `ricos-shortcuts`
+
   - [#3835](https://github.com/wix/ricos/pull/3835) react-hotkeys works with tiptap
   - [#3841](https://github.com/wix/ricos/pull/3841) fix keys